core0: Calculate positional vector coordinates through LSM6DSOX
In order to figure out what is physically going on with the robot use
the accelerometer to figure out the direction in which we are moving
and the gyroscope to know when the enemy would lift us up.

An `acc_Ry_error` was used since, empirically, we found out that there
is an error rate between what we expected and what we received on the
OY axis of the accelerometer.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>